backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
langchain
langchain-community
langchain-groq
langchain-text-splitters
fastembed
chromadb
pydantic
python-multipart
python-dotenv
pysqlite3-binary
